support custom op in inference
